On XFS the file permissions appear to apply also to the extended attribute data. ie. If a file is not readable then nor is the extended attribute data - even if there isn't any this is still an error.
This is not an ideal fix for this problem. Really we should notify a read-error as we do for file read permission problems and continue, but the BackupClientFileAttributes code is called from multiple locations and it is not clear that it could always report the error in such a way, nor even continue. This alternative is better than aborting the whole run. In the case where a file isn't readable this will be found and reported elsewhere anyway.