Add --bbackupd-args and --bbstored-args options, which may be used to pass additional command-line parameters to bbackupd and bbstored in some tests.
Add -t and -T options, which add tags and timestamps to log messages using
the logging framework inside the tests themselves.

Now you can do things like:

	./t -Tt test \
		--bbackupd-args="-VTkt bbackupd" \
		--bbstored-args="-VTkt bbstored"

and you will get output from the test itself, the bbackupd and bbstored
daemons, all interleaved, timestamped and tagged on the console.

This is very useful for debugging synchronisation problems, where 
interleaved output from bbackupd and bbstored in the system logs is not 
enough, because you want to see exactly how they interact with the test,
or you don't have access to the system logs.

(refs #3)