Use @sysconfdir_expanded@ as default location for configuration files, instead of /etc/box. Thanks to Eric Cronin for pointing this out.
Likewise, use @localstatedir_expanded@ as default location for temporary
state (*.pid and bbackupd.sock), defaults to /usr/local/var, although
/var/run may be better on LFS systems.

Likewise, use @bindir_expanded@ for binaries.

Reformat usage text for readability.

Remove the message that "Directories not checked against mountpoints",
as they now are.

Report the type of the unrecognised message in the "unknown error" part
of the default NotifyScript.


git-svn-id: https://www.boxbackup.org/svn/box/trunk@2063 9c52e268-1103-0410-960d-c6a436f6b474