Track and log file deletions by name.
Split crypto init and file sync process into its own method, to reduce 
call depth and facilitate calling in process from tests.

Differentiate between 3 uses of stat in BackupClientDirectoryRecord by 
renaming the structures.

Use stat instead of lstat when checking the filesystem that's holding an 
entity, in case it's a symbolic link to a different filesystem.