Don't read the whole uploaded body in HTTPRequest::Receive, as the client may be expecting a 100 Continue header (or other response) before sending it, and only the HTTPServer should send that for us.
Keep track of the stream that we're reading from, in case there's a body 
to read later.

Simplify parsing of HTTP method, and add support for PUT requests.

Add support for parsing Expect headers and storing and retrieving any
unrecognised headers.

Add support for sending a streaming upload from an IOStream with an HTTP 
request as the body (e.g. for PUT requests).


git-svn-id: https://www.boxbackup.org/svn/box/trunk@2438 9c52e268-1103-0410-960d-c6a436f6b474