Improve debugging messages after repairing a directory.
Report whether the directory still had errors after the first pass of
dir.CheckAndFix(), and also whether a file is marked as both Old and
Deleted, which means that its block usage is not accounted for anywhere.

There's possibly still a bug here where an invalid directory entry that
is itself a directory may not result in the parent directory being marked
as isModified and written back out afterwards, which may be what's
happening to Markus Grunwald (failure to correct the error with "check fix").

git-svn-id: https://www.boxbackup.org/svn/box/trunk@3178 9c52e268-1103-0410-960d-c6a436f6b474