Fix getopt header confusion.
Our copy of getopt.h in lib/win32 was beating the system getopt.h in the
search path, and providing different definitions, which clang picked up.

Renamed our getopt.h to bsd_getopt.h, and added a box_getopt.h to decide
whether to include this one instead of the system getopt.h.

git-svn-id: https://www.boxbackup.org/svn/box/trunk@3198 9c52e268-1103-0410-960d-c6a436f6b474