Create new refcount database atomically during account check.
Use a temporary refcount db for check instead of an in-memory vector. This
avoid the memory usage problems created by using the vector on large accounts,
but may require us to improve the efficiency of the refcount database itself
to avoid large numbers of small I/O operations. That is very doable now that
we're using a class for it.

Fix some inconsistencies and mistakes in handling reference counts and
info counters during account checks (more to come).

git-svn-id: https://www.boxbackup.org/svn/box/trunk@3229 9c52e268-1103-0410-960d-c6a436f6b474