Fix compile error.
Caused by removal of BackupStoreRefCountDatabase::CreateForRegeneration.
It's not safe to do this in housekeeping anyway, because it might not
complete the scan, leaving an incomplete database, that might persuade
BackupStoreContext that it's OK to delete an object that still really has
references to it.

If your refcount db disappears or is damaged, you need to run a check fix.