Fix an exception checking refcounts in housekeeping.
If the old refcount database was shorter than the new one, then we could end
up trying to access a nonexistent refcount, triggering an exception, which
prevented all comparison of the old and new refcount databases.

Also, from now on, a mismatch between old and new refcounts is treated as a
housekeeping error, which is detectable in tests.