Directories' entries in parent directory should track current size.
When entries are added to a directory by a command (BackupStoreContext),
and when entries are removed from a directory (by housekeeping), update the
parent directory's entry for us if our size has changed.

Make BackupStoreCheck check for, report and fix errors when directory entries
are directories and the size is wrong (as well as files).

Conflicts:

	test/backupstore/testbackupstore.cpp

Fix directories loaded without size being set, leading to warnings later.

We can't check that the old size in the parent entry matched the old real size
of the directory, unless we set the old real size in the directory.

And we don't need to pass the old directory size to
HousekeepStoreAccount::DeleteFile, because we can get it from the directory
itself.