Backport Timers::Cleanup that's safe to use in test cleanup.
Allows it not to throw an exception if timers weren't initialised when
cleanup was requested. Normally we want an exception thrown, but not while
we're cleaning up a test that might have failed with timers uninitialised.

More timers fixes after cleanup no-exception option.

Merged back changes from the test refactor branch to reduce diffs.