Allow access to BackupDaemon's BackupClientContext for tests.
The std::auto_ptr is now returned unclosed. If the caller of RunSyncNow()
does nothing with the return value, then it will be released, closing the
connection. However the caller can save the std::auto_ptr locally and
poke around in it if necessary. This interface is designed for tests only!

Merged back changes from the test refactor branch to reduce diffs.

Fix failure to record client store marker, after BackupClientContext refactor.

This caused the daemon not to abort safely if the CSM changed under our feet,
meaning that our cached data is invalid.

Fix for earlier patch to allow tests to access the BackupClientContext after
backup completes, because a new CSM wasn't generated in time for bbackupd to
record it, because the connection wasn't closed yet.