Handle raidfile versioning better on low time accuracy systems.
We simulate a 1 second accuracy clock in DEBUG mode on all systems, to help
catch corner cases in tests. This now applies to both RAID and non-RAID files.

We always add file size to the version number to help detect changes,
especially in such cases. In RELEASE builds, we use as much accuracy from the
clock as we can get, as well as file size.

git-svn-id: https://www.boxbackup.org/svn/box/trunk@3390 9c52e268-1103-0410-960d-c6a436f6b474