Fix error in BufferedStream::StreamDataLeft.
StreamDataLeft() incorrectly returned false when there was still data
buffered, which made IOStream::CopyStreamTo stop copying early, resulting
in incomplete files when copying data out of a BufferedStream. This meant
that tests involving a local protocol, where a BufferedStream was copied
directly to a RaidFile, resulted in corrupted files that would not verify.

git-svn-id: https://www.boxbackup.org/svn/box/trunk@3395 9c52e268-1103-0410-960d-c6a436f6b474