In debug builds, invalidate cache items instead of deleting them.
This allows us to trigger an assertion at the moment that an invalid item
is used, instead of memory corruption, segfaults and other weird behavior
such as directories mysteriously changing identity because they're invalid.

This allows debug builds to strictly and accurately enforce correct (valid)
use of cached directory objects in BackupStoreContext.