Refactor handling of exceptions in protocol server command executors.
Add a standard method to Replyable that will be called if a recoverable
exception (a BoxException) occurs, and can return a protocol Message to be
sent to the client, such as an error code for various standard errors, or
rethrow the exception.

If you want something different, catch exceptions and return the desired
reply yourself, or you'll get the default handling.