Remove all timing dependency from test_ssl_keepalives().
This test has always been fragile, and usually fails on Travis for reasons
unknown, so replace the log parsing and intercepts with simple mocking code.
This doesn't cover 100% of what the old test did, including diff timer aborts
and the number of blocks used by the uploaded files, but it should be
completely robust.