Fix test/backupstorefix random failures caused by file upload order.
Previously, files were uploaded using bbackupd, which did it in the order that
it read the directories, which was OS-dependent. So destroying certain fixed
RaidFiles would have an unpredictable effect on the store. Now, we write a manifest
when creating the test files, and upload them in exactly that order, so that the
ID of each file will always be the same.