Create a job object during Windows tests to kill abandoned daemons
Should prevent some cases of tests hanging on AppVeyor due to child
process daemons that were started and not stopped.

Delete files between tests using much more similar code (and the same
list) on Windows and Linux, to reduce risk of divergence causing tests
to pass on one and fail on the other.