Improve detection of warnings not supported by compilers.
GCC normally only gives a warning on unsupported warnings, so our M4 feature
detection can think that a warning is supported, and enable it, when in fact
it's not supported, and this adds harmful warning noise to the compilation
output. By making -Wunknown-warning-option (enabled by default) into a fatal
error when checking whether warnings are supported, this allows us to really
detect whether they are or not.