Fix build on SmartOS by switching back to GNU Make
I have no idea where I got the idea that there was a "bmake" command on SmartOS
that is BSD make, or why that would be better than gmake (GNU make). I can't
find any reference to the existence of such a thing. So I've just switched back
to using gmake unconditionally.