Hopefully fix test/backupstorefix slow performance on Windows
Frequent directory writes, combined with NtClose() being very slow, meant that
the test took 30-120 seconds on Windows (compared to about 5-10 seconds on OSX)
and frequently timed out on Travis.

I've implemented a write cache for directories in BackupStoreContext, and made
the read cache effective in DEBUG builds, to hopefully address this, at the
cost of some extra complexity in tests with multiple sessions.