Improve directory cache efficiency with S3 servers
We used to always hit the S3 server to check whether our cached
BackupStoreDirectory was stale.  But if we are holding a write lock, then in
theory nobody should be changing these objects except us, so the cache
shouldn't go stale, so we shouldn't need to check and can avoid hitting the
server at all.