CMake builds: revert to install(PROGRAMS)
The change to install(TARGETS) meant that application binaries such as
bbstored.exe and s3simulator.exe were installed with the wrong names,
and there seems to be no way to built executables with different names
than their add_executable target names (which must be unique).

But at least reverting to install(PROGRAMS) allows us to rename them to
the correct final names on installation.