Build Win32/64 binary packages automatically
Every AppVeyor build on the specified branches that passes will be packaged
into a ZIP file and installer using CPack and uploaded to GitHub Releases.

Generate version number entirely in AppVeyor config, and build it into the
binary, and use it in the names of the created binary packages, and compile it
into the binaries.

(cherry picked from commit c5d8173bf03ab9fa93d3b3ffd871b378b0819567)