Fix loss of stream sync when a local I/O error occurs during compare
This would cause all subsequent file compares to fail with Protocol_ObjTooBig
errors.  Fixed by wrapping the stream returned by BackupProtocolServer and
Client's ReceiveStream method in a SelfFlushingStream, so that it will always
be fully read (flushed) if either server or client throws an exception while
processing the request/response.