Clear up Daemon command-line arguments -D and -F
By separating their roles (which used to overlap) it's now possible to run a
forking server in the foreground, or a non-forking daemon in the background,
which can be useful for debugging (for example, to more closely simulate how
test daemons work on Windows when running on Unix). -DF is now needed to
have a non-forking server in the foreground.