Fix ProtocolError single-argument (subtype only) constructor
Correctly initialise mType from the ErrorType CONSTANT in the protocol message
definition in this case, instead of leaving it uninitialised. Add test. Also
test that these unexpected errors cause the connection to be closed as
expected.

Also catch SIGPIPE so that the test doesn't die when the server process closes
the connection to it.