Remove MEMLEAKFINDER_NOT_A_LEAK from FileStream
It's not clear why we need to tell the memory leak tracking system to ignore
FileStream objects if they raise an exception in their constructors. The
objects should still be destroyed normally. Additionally, now that we
allocate more of them on the heap (for NamedLock), this can cause later
allocations to be silently ignored as not-leaks, which is causing
test_memory_leak_detection to fail.

(cherry picked from commit 24386d968c3fd9a9c486e42c24e2844e9ff713ad)