Add support for patch combining in S3 stores
This allows housekeeping to reduce the storage required. It's very inefficient,
but needed for compatibility with bbstored stores.

Convert test/backupstore test_server_housekeeping and test/backupstorepatch to
specialised tests, and make them test S3 stores as well. Change the
GetBlockIndexByName command to reconstruct the block index when the most recent
version of the file is a patch (as it will be on S3 stores).

Stop pretending that S3BackupFileSystem cannot merge patches, and stop
reporting that it's "hard" to do so for the time being, at least until we have
a better way to manage store space (e.g. snapshots).