Fix "timeouts" in test/basicserver on FreeBSD
If the death of the server was not detected while sending the Quit message to
it, but only later when trying to read the reply (which seems only to happen on
FreeBSD), the empty read from the dead socket caused ReadFullBuffer to return
false, and this was mis-interpreted as a timeout.

Now SocketStream::ReadFullBuffer overrides IOStream::ReadFullBuffer to check
the socket status when this happens, and if the socket is closed then it raises
a more appropriate exception.  test/basicserver now catches this exception too.