Cache compiled packages to speed up unibuilds
This is particularly relevant to Windows builds, where ccache is not available,
so zlib, openssl and pcre must be compiled from scratch on every build.  This
process caches the installed files (in the install/ directory) and rebuilds
them if the CMakeLists file changes.

The cache is cleared on appveyor if CMakeLists.txt changes (which invalidates
the old packages anyway) but not on Travis, which may grow to become an issue
in future.