Add extra checks that daemons are really ready to serve in tests
Removing some sleeps has exposed race conditions where daemons write a PID file
before they are actually ready to accept connections, and sometimes we would
hit that in tests. Now we pass a port or UNIX socket (or both) to LaunchServer,
and it waits for them to start accepting connections before returning.