WaitForServerStartup: fix disturbance due to connecting to sockets
Attempting a connection to the daemon's server socket seems to upset it, making
it unable to accept connections for some seconds afterwards, which defeats the
point of the test.

Instead, we have modified the server (Daemon) to not finish writing the PID
file until it is completely ready to serve requests.