Loosen timing tolerance by reducing MaxUploadWait
Before, we only had 1 second tolerance (we stopped updating the file 1 second
before bbackupd would run to back it up) and if bbackupd ran at the wrong time
(just before a second boundary) and lost time so that it went over that second
boundary, then it would backup the file one run earlier than the test expected.

Increasing the tolerance by 1 second should make that really unlikely.

Also increase some INFO log messages to NOTICE, because they (and their
timestamps) are essential to debugging the test when it fails, even in release
mode. And 4 extra lines of test output is not a high price to pay.

Also stop "adjusting" MaxUploadWait so that it actually means what it appears
to mean: the time between when a file was first seen, and when it will be
uploaded, even if it continues to be modified in the mean time.