Split test/backupstorefix into individual test cases
Fix block, file and diretory count and refcount accounting during store repair,
to avoid leaving the store with further errors that cause test failures.

Refactor out counting of files/blocks by directory entry, and repairing
individual unattached objects, into their own functions.

Correctly set the refcount for a lost+found directory when it's created (or
should be created) on the store, to ensure that the test fails if it's not.