Change emu_rename semantics to allow renaming over existing files
Remove explicit deletion before rename in
BackupStoreRefCountDatabase::Commit. It seems that this sometimes isn't
atomic for release builds on Windows, or at least it would explain the
random test failures on AppVeyor. The new semantics better match UNIX,
and the improved emulation should allow the removal of plenty of
conditional compatibility code.