Hopefully fix issue with renaming over recently-closed files on Windows
This caused random test failures, especially on AppVeyor in test/backupstore
`test_cannot_open_multiple_writable_connections(store)`, which looked like this:

> [00:12:36] 6: [check fix] WARNING: Exception thrown: CommonException(OSFileError) (Failed to rename temporary refcount database file from testfiles/0_0\backup\01234567\refcount.rdb.rfwX to testfiles/0_0\backup\01234567\refcount.rdb.rfw: Access is denied. (5)) at C:\projects\boxbackup\lib\backupstore\BackupStoreRefCountDatabase.cpp:200

This is evil:

> Even though you've called CloseHandle on the file handle, the
> kernel may still have outstanding references that take a few milliseconds to
> close... Windows is notorious for this issue. sqlite handles the problem by
> retrying the delete operation every 100 milliseconds up to a maximum number.

https://stackoverflow.com/questions/1753209/deletefile-fails-on-recently-closed-file

Note that I've only fixed the issue for rename-over, not deletion. The only
race condition involving deletion that I'm aware of was just before
rename-over, which is no longer required now that we use MoveFileEx, so should
be removed anyway.