BackupStoreRefCountDatabase on S3: truncate existing file before overwriting
Otherwise we may download a smaller file from the server, leaving existing
stale data at the end of the old file, and thus fail to match the MD5 checksum
after downloading.

Use a TeeStream to calculate the MD5 checksum while streaming the file, without
needing to read it back again.