FileStream: implement real file locking on Windows
Previously we just used dwShareMode, but that doesn't give us a way
to wait for a lock, while LockFileEx does. Unfortunately LockFileEx
is mandatory locking, so you can't write to a file that is open in
SHARED mode, so change the default locking mode to NONE for
compatibility, without having to add EXCLUSIVE locks to all files
opened for writing.