BackupStoreCheck: fix failure to compare refcounts of last object ID
Fix test that would randomly detect an extra change to the refcount of an
object, depending if the test broke the object with the highest unused object
ID (which depended on upload order).

Add ability to ignore changes to a specific object ID, and use it to ignore
changes to any newly-created lost+found directory, as these are expected.