ServerControl: add signal checks and wait() wrapper
KillServer: add a parameter to allow the expected exit signal to be checked.
This allows us to determine if a process died for the wrong reason, e.g. if it
failed to catch SIGTERM, or was killed by SIGSEGV instead.

WaitForProcessExit: split out, to make it easier to clean up after forked
daemons, and check their exit status and signal.

test/bbackupd: use these to improve handling and failure mode checking of
forking daemon tests, reducing risk of hangs.