Cirrus: Add chroot support and Debian i386 build
The Debian project requires both amd64 and i386 builds to pass. In
Buildbot we accomplished this by creating an i386 chroot on an amd64 EC2
instance and running the buildbot daemon inside that.

This tries to accomplish similar, but as Cirrus doesn't run a daemon, we
need to run many chroot commands. Debian's schroot allows setting up a
persistent environment with things like /tmp bind mounted from the host.
This is useful as Cirrus uses a directory in /tmp as its working dir.