Fix for overlapping locations with VSS
Kai Liebenau reports: "I have found a little issue in the windows version of
daemon. The issue is the working with snapshots, when you already backup more
then one locations from the same volume. In this case, is the whole vole will
be backuped in the second location."

This is his patch, untested.