Daemon: always block when reopening PID file, and check for lost race instead
Even if we avoid blocking, the race window is nonzero, and if the daemon forks
a child process (e.g. bbstored housekeeping) then we need to wait for it to
close the exclusively-locked file anyway, and on some systems (e.g. Travis CI)
we still can't reliably grab the shared lock immediately. So we always block,
but check after acquiring the shared lock that the PID has not changed, to
ensure that we detect and abort if we lost the race.