Change uid before modifying account in bbstoreaccounts
bbstored is often run as root, and changes uid to the specified user. In such
cases, bbstoreaccounts is probably run as root too, but did not change uid, so
it would change the ownership of files in the store during "check fix", making
the account unusable until the permissions were fix.

Fix this by changing uid in bbstoreaccounts too, if configured in bbstored.

Thanks to Jack Warkentin for the report.