Fix handling of SSLSecurityLevel
By default we now (4 years after the original issue) keep the
system-wide SSL security level. We leave users with the option to set
SSLSecurityLevel manually to override it.

The previous behaviour was quite broken: we documented a special meaning
for level 0, but in fact this is a standard level defined by OpenSSL.
And since they downgraded the security of SHA1 MD, so that it's only
accepted at level 0, we now need to use that level for our old
certificates to work.

We also failed to actually leave the system-wide SSL security level
alone when SSLSecurityLevel was set to BOX_DEFAULT_SSL_SECURITY_LEVEL
(-1), but now we do.

We don't have any certificates which are (still) valid at level 1 but
not level 2 to test with, so remove the level 1 tests. Since we test
levels 0 and 2, it seems pretty pointless to also test level 1, so not
worth the effort and complexity to try to craft new certificates to do
so.

We failed to set SSLSecurityLevel = 0 in many test configuration files
for basicserver daemones, bbackupd and bbstored, so they would never
have worked. I don't know how this passed initial testing. It definitely
fails on Ubuntu 22.10.